Added hard-core repulsion into the event generator in GUI, EV Crossterms now uses VDW-HRG model implementation with a = 0
