+++ CMakeLists.txt
@@ -8,7 +8,7 @@ project (ThermalFIST)
 # The version number.
 set (ThermalFIST_VERSION_MAJOR 1)
 set (ThermalFIST_VERSION_MINOR 3)
-set (ThermalFIST_VERSION_DEVEL 4)
+set (ThermalFIST_VERSION_DEVEL 5)
 
 # configure a header file to pass some of the CMake settings
 # to the source code
